Update and rename Integrated Script.py to serper_search.py
